LXQt Now!
The thing is, having something modern is not guaranteed to be better at things. COSMIC is modern, but not better in terms of performance and efficiency. LXQt consumes less resources, is suitable for my potato PC and opens more memory for other programs. Therefore, LXQt is more productive. This might change in the future. It supports Wayland a modern protocol, but truly better than Xorg, as it is not only modern, but has a modern(better) approach.